Validate start time is at least 1 minute in the future
