+++ Loop/Views/Presets/CreatePresetNameAndScheduledEdit.swift
@@ -137,10 +137,13 @@ struct CreatePresetNameAndScheduledEdit: View {
                         Toggle("", isOn: Binding(get: {
                             return preset.startDate != nil
                         }, set: { newValue in
-                            if newValue {
-                                preset.startDate = Date()
-                            } else {
-                                preset.startDate = nil
+                            withAnimation {
+                                if newValue {
+                                    preset.startDate = Date().addingTimeInterval(.hours(1))
+                                } else {
+                                    preset.startDate = nil
+                                    preset.repeatOptions = nil
+                                }
                             }
                         }))
                         .toggleStyle(SwitchToggleStyle(tint: .green))
@@ -173,7 +176,7 @@ struct CreatePresetNameAndScheduledEdit: View {
                         HStack {
                             Text("Repeat")
                             Spacer()
-                            Picker("Repeat", selection: $selectedRepeatOption) {
+                            Picker("Repeat", selection: $selectedRepeatOption.animation()) {
                                 ForEach(RepeatOption.allCases, id: \.self) { option in
                                     Text(String(describing: option))
                                 }
@@ -187,29 +190,29 @@ struct CreatePresetNameAndScheduledEdit: View {
                             Divider()
                                 .padding(.top, -4)
                             HStack {
-                                Button(action: {
-                                    showingDayPicker = true
-                                }) {
-                                    HStack {
-                                        Text("Selected days")
-                                            .foregroundColor(.primary)
-                                        Spacer()
-                                        RepeatOptionView(repeatOptions: preset.repeatOptions ?? .none)
-                                    }
-                                    .padding(.vertical, 6)
+                                Text("Selected days")
+                                    .foregroundColor(.primary)
+                                HStack {
+                                    Spacer()
+                                    RepeatOptionView(repeatOptions: preset.repeatOptions ?? .none)
+                                        .padding(.vertical, 6)
+                                        .onTapGesture {
+                                            withAnimation {
+                                                showingDayPicker = true
+                                            }
+                                        }
                                 }
                                 .popover(isPresented: $showingDayPicker) {
                                     DayPickerPopup(selectedDays: Binding(
-                                    get: {
-                                        preset.repeatOptions ?? .none
-                                    }, set: { newValue in
-                                        preset.repeatOptions = newValue.union(requiredRepeatOption ?? .none)
-                                    }))
+                                        get: {
+                                            preset.repeatOptions ?? .none
+                                        }, set: { newValue in
+                                            preset.repeatOptions = newValue.union(requiredRepeatOption ?? .none)
+                                        }))
                                     .cornerRadius(12)
                                     .presentationCompactAdaptation(.popover)
                                 }
                             }
-
                         }
                     }
                 }
@@ -265,95 +268,6 @@ struct CreatePresetNameAndScheduledEdit: View {
 
     var allowSave: Bool {
         return (!preset.savePreset && preset.duration != nil) || (preset.savePreset && !preset.name.isEmpty && preset.duration != nil)
-    }
-}
-
-// Optional: Extension for expanded row when Duration is tapped
-struct DurationRowView: View {
-    @Binding var untilTurnOff: Bool
-    @Binding var selectedHour: Int
-    @Binding var selectedMinute: Int
-
-    let hours = Array(0...23)
-    let minutes = [0, 5, 10, 15, 20, 25, 30, 35, 40, 45, 50, 55]
-
-    var body: some View {
-        VStack {
-            HStack(spacing: 0) {
-                Picker("Hour", selection: $selectedHour) {
-                    ForEach(hours, id: \.self) { hour in
-                        Text("\(hour)").tag(hour)
-                    }
-                }
-                .pickerStyle(WheelPickerStyle())
-                .frame(width: UIScreen.main.bounds.width / 2)
-                .clipped()
-
-                Text("hour")
-                    .foregroundColor(.gray)
-                    .padding(.trailing, 20)
-
-                Picker("Minute", selection: $selectedMinute) {
-                    ForEach(minutes, id: \.self) { minute in
-                        Text("\(minute)").tag(minute)
-                    }
-                }
-                .pickerStyle(WheelPickerStyle())
-                .frame(width: UIScreen.main.bounds.width / 2)
-                .clipped()
-
-                Text("min")
-                    .foregroundColor(.gray)
-            }
-            .frame(height: 120)
-
-            HStack {
-                Text("Until I turn off")
-                    .font(.body)
-
-                Spacer()
-
-                Toggle("", isOn: $untilTurnOff)
-                    .toggleStyle(SwitchToggleStyle(tint: .green))
-                    .labelsHidden()
-            }
-            .padding(.horizontal, 16)
-        }
-    }
-}
-
-struct RepeatOptionView: View {
-    let repeatOptions: PresetScheduleRepeatOptions
-
-    private var selectedDays: [PresetScheduleRepeatOptions] {
-        PresetScheduleRepeatOptions.allCases.filter { repeatOptions.contains($0) }
-    }
-
-    private var isSingleDay: Bool {
-        selectedDays.count == 1
-    }
-
-    var body: some View {
-        if repeatOptions == .none {
-            Text(repeatOptions.description)
-                .tint(.secondary)
-        } else if isSingleDay {
-            Text(selectedDays[0].description)
-                .foregroundColor(.secondary)
-        } else {
-            HStack(spacing: 4) {
-                ForEach(PresetScheduleRepeatOptions.allCases, id: \.rawValue) { day in
-                    Text(String(day.description.first!))
-                        .font(.system(size: 12))
-                        .frame(width: 20, height: 20)
-                        .background(
-                            Circle()
-                                .fill(repeatOptions.contains(day) ? Color.blue : Color.gray.opacity(0.2))
-                        )
-                        .foregroundColor(repeatOptions.contains(day) ? .white : .gray)
-                }
-            }
-        }
     }
 }
